+++ Assets/TiedanSouls/Template/TM/Generic/Model/AttributeEffectTM.cs
@@ -11,19 +11,19 @@ namespace TiedanSouls.Template {
     public struct AttributeEffectTM {
 
         [Header("当前生命值 数值计算类型")] public NumCalculationType hpNCT;
-        [Header("当前生命值 影响值")] public float hpEV;
+        [Header("当前生命值 影响值")] public int hpEV;
 
         [Header("最大生命值 数值计算类型")] public NumCalculationType maxHPNCT;
-        [Header("最大生命值 影响值")] public float maxHPEV;
+        [Header("最大生命值 影响值")] public int maxHPEV;
 
         [Header("攻击力 数值计算类型")] public NumCalculationType atkPowerNCT;
-        [Header("攻击力影响值")] public float atkPowerEV;
+        [Header("攻击力影响值")] public int atkPowerEV;
 
         [Header("攻击速度 数值计算类型")] public NumCalculationType atkSpeedNCT;
-        [Header("攻击速度 影响值")] public float atkSpeedEV;
+        [Header("攻击速度 影响值")] public int atkSpeedEV;
 
         [Header("移动速度 数值计算类型")] public NumCalculationType moveSpeedNCT;
-        [Header("移动速度 影响值")] public float moveSpeedEV;
+        [Header("移动速度 影响值")] public int moveSpeedEV;
 
     }
